requirements.txt
amqp==5.0.6
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
gevent==21.8.0
greenlet==1.1.2
grequests==0.6.0
idna==3.3
kombu==5.2.2
prompt-toolkit==3.0.24
pymongo==4.0.1
python-decouple==3.5
pytz==2021.3
redis==3.5.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
zope.event==4.5.0
zope.interface==5.4.0
